Move reading routines into _serializers.read subpackage
